Trait rustcoalescence_algorithms_cuda_kernel::Kernel

pub unsafe trait Kernel<M: MathsCore, H: Habitat<M> + RustToCuda, G: PrimeableRng<M> + RustToCuda, R: LineageReference<M, H>, S: LineageStore<M, H, R> + RustToCuda, X: EmigrationExit<M, H, G, R, S> + RustToCuda, D: DispersalSampler<M, H, G> + RustToCuda, C: CoalescenceSampler<M, H, R, S> + RustToCuda, T: TurnoverRate<M, H> + RustToCuda, N: SpeciationProbability<M, H> + RustToCuda, E: MinSpeciationTrackingEventSampler<M, H, G, R, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry<M> + RustToCuda, A: SingularActiveLineageSampler<M, H, G, R, S, X, D, C, T, N, E, I> + RustToCuda, ReportSpeciation: Boolean, ReportDispersal: Boolean> {
    fn get_ptx_str() -> &'static str
    where
        Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>;

    fn new_kernel(
    ) -> CudaResult<TypedKernel<dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>>
    where
        Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>;

    fn simulate<'__r2c_lt_0, '__r2c_lt_1, '__r2c_lt_2, '__r2c_lt_3, '__r2c_lt_4, '__r2c_lt_5, '__r2c_lt_6>(
        &mut self, 
        simulation: &'__r2c_lt_0 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_0, 
        task_list: &'__r2c_lt_1 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_1, 
        event_buffer_reporter: &'__r2c_lt_2 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_2, 
        min_spec_sample_buffer: &'__r2c_lt_3 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_3, 
        next_event_time_buffer: &'__r2c_lt_4 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_4, 
        total_time_max: &'__r2c_lt_5 <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_5, 
        total_steps_sum: &'__r2c_lt_6 <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_6, 
        max_steps: <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_7, 
        max_next_event_time: <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_8
    ) -> CudaResult<()>
    where
        Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>;

    fn simulate_raw<'__r2c_lt_0, '__r2c_lt_1, '__r2c_lt_2, '__r2c_lt_3, '__r2c_lt_4, '__r2c_lt_5, '__r2c_lt_6>(
        &mut self, 
        simulation: HostAndDeviceMutRef<'__r2c_lt_0, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_0 as RustToCuda>::CudaRepresentation>>, 
        task_list: HostAndDeviceMutRef<'__r2c_lt_1, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_1 as RustToCuda>::CudaRepresentation>>, 
        event_buffer_reporter: HostAndDeviceMutRef<'__r2c_lt_2, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_2 as RustToCuda>::CudaRepresentation>>, 
        min_spec_sample_buffer: HostAndDeviceMutRef<'__r2c_lt_3, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_3 as RustToCuda>::CudaRepresentation>>, 
        next_event_time_buffer: HostAndDeviceMutRef<'__r2c_lt_4, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_4 as RustToCuda>::CudaRepresentation>>, 
        total_time_max: HostAndDeviceConstRef<'__r2c_lt_5, SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_5>>, 
        total_steps_sum: HostAndDeviceConstRef<'__r2c_lt_6, SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_6>>, 
        max_steps: SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_7>, 
        max_next_event_time: SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_8>
    ) -> CudaResult<()>
    where
        Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>;
}

Required methods

fn get_ptx_str() -> &'static str where
Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>,

fn new_kernel(
) -> CudaResult<TypedKernel<dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>> where
Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>,

fn simulate<'__r2c_lt_0, '__r2c_lt_1, '__r2c_lt_2, '__r2c_lt_3, '__r2c_lt_4, '__r2c_lt_5, '__r2c_lt_6>(
&mut self,
simulation: &'__r2c_lt_0 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_0,
task_list: &'__r2c_lt_1 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_1,
event_buffer_reporter: &'__r2c_lt_2 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_2,
min_spec_sample_buffer: &'__r2c_lt_3 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_3,
next_event_time_buffer: &'__r2c_lt_4 mut <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_4,
total_time_max: &'__r2c_lt_5 <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_5,
total_steps_sum: &'__r2c_lt_6 <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_6,
max_steps: <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_7,
max_next_event_time: <() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_8
) -> CudaResult<()> where
Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>,

fn simulate_raw<'__r2c_lt_0, '__r2c_lt_1, '__r2c_lt_2, '__r2c_lt_3, '__r2c_lt_4, '__r2c_lt_5, '__r2c_lt_6>(
&mut self,
simulation: HostAndDeviceMutRef<'__r2c_lt_0, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_0 as RustToCuda>::CudaRepresentation>>,
task_list: HostAndDeviceMutRef<'__r2c_lt_1, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_1 as RustToCuda>::CudaRepresentation>>,
event_buffer_reporter: HostAndDeviceMutRef<'__r2c_lt_2, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_2 as RustToCuda>::CudaRepresentation>>,
min_spec_sample_buffer: HostAndDeviceMutRef<'__r2c_lt_3, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_3 as RustToCuda>::CudaRepresentation>>,
next_event_time_buffer: HostAndDeviceMutRef<'__r2c_lt_4, DeviceAccessible<<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_4 as RustToCuda>::CudaRepresentation>>,
total_time_max: HostAndDeviceConstRef<'__r2c_lt_5, SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_5>>,
total_steps_sum: HostAndDeviceConstRef<'__r2c_lt_6, SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_6>>,
max_steps: SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_7>,
max_next_event_time: SafeDeviceCopyWrapper<<() as KernelArgs<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>::__T_8>
) -> CudaResult<()> where
Self: Sized + Launcher<KernelTraitObject = dyn Kernel<M, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>>,